py/requirements.txt
gymnasium[all]
pyglet
py-color-log
